consolidate server files
