+++ hyperdome_server/web/web.py
@@ -24,6 +24,10 @@ import os
 import queue
 import socket
 from urllib.request import urlopen
+import random
+import binascii
+import traceback
+import json
 
 import flask
 from flask import Flask, request, render_template, abort, make_response
@@ -73,8 +77,6 @@ class Web(object):
         if self.common.debug:
             self.debug_mode()
 
-        # Are we running in GUI mode?
-        self.is_gui = is_gui
 
         # If the user stops the server while a transfer is in progress, it
         # should immediately stop the transfer. In order to make it
@@ -82,7 +84,6 @@ class Web(object):
         # then the user stopped the server.
         self.stop_q = queue.Queue()
 
-        self.mode = 'share'
 
         self.security_headers = [
             ('Content-Security-Policy',
@@ -92,7 +93,7 @@ class Web(object):
             ('X-Xss-Protection', '1; mode=block'),
             ('X-Content-Type-Options', 'nosniff'),
             ('Referrer-Policy', 'no-referrer'),
-            ('Server', 'OnionShare')]
+            ('Server', 'Hyperdome')]
 
         self.q = queue.Queue()
         self.slug = None
@@ -110,8 +111,17 @@ class Web(object):
         # Define the web app routes
         self.define_common_routes()
 
-        # Create the mode web object, which defines its own routes
-        self.share_mode = ShareModeWeb(self.common, self)
+        # hyperdome server user tracking variables
+        self.counselors_available = dict()
+        self.active_chat_user_map = dict()
+        self.pending_messages = dict()
+        self.guest_keys = dict()
+        self.counselor_keys = dict()
+
+        #
+        self.info = {'name': 'hyperdome',
+                     'version': self.common.version,
+                     'online': str(len(self.counselors_available))}
 
     def define_common_routes(self):
         """
@@ -141,6 +151,110 @@ class Web(object):
             """
             r = make_response(render_template('receive_noscript_xss.html'))
             return self.add_security_headers(r)
+
+        @self.app.errorhandler(Exception)
+        def unhandled_exception(e):
+            e_str = ''.join(traceback.format_exception(type(e),
+                                                       e,
+                                                       e.__traceback__))
+            print(e_str)
+            return "Exception raised", 500
+
+        @self.app.route("/probe")
+        def probe():
+            self.info['online'] = str(len(self.counselors_available))
+            return json.dumps(self.info)
+
+        @self.app.route("/request_counselor", methods=['POST'])
+        def request_counselor():
+            guest_id = request.form['guest_id']
+            guest_key = request.form['pub_key']
+            counselors = [
+                counselor for counselor, capacity in self.counselors_available.items() if capacity]
+            if not counselors:
+                return ''
+            chosen_counselor = random.choice(counselors)
+            self.counselors_available[chosen_counselor] -= 1
+            self.active_chat_user_map[guest_id] = chosen_counselor
+            self.active_chat_user_map[chosen_counselor] = guest_id
+            self.guest_keys[chosen_counselor] = guest_key
+            counselor_key = self.counselor_keys.pop(chosen_counselor)
+            return counselor_key
+
+        @self.app.route("/poll_connected_guest", methods=['GET'])
+        def poll_connected_guest():
+            counselor_id = request.form['counselor_id']
+            guest_key = self.guest_keys.pop(counselor_id, '')
+            return guest_key
+
+        @self.app.route("/counseling_complete", methods=['POST'])
+        def counseling_complete():
+            sid = request.form['user_id']
+            if sid not in self.active_chat_user_map:
+                return 'no active chat', 404
+            other_user = self.active_chat_user_map[sid]
+            self.active_chat_user_map.pop(sid)
+            self.pending_messages.pop(sid, '')
+            self.active_chat_user_map.update({other_user: ''})
+            if sid in self.counselors_available:
+                counselor_id = sid
+            elif other_user in self.counselors_available:
+                counselor_id = other_user
+            else:
+                return 'Counselor has left the chat'
+            self.counselors_available[counselor_id] += 1
+            return 'Chat Ended'
+
+        @self.app.route("/counselor_signout", methods=["POST"])
+        def counselor_signout():
+            sid = request.form['user_id']
+            self.counselors_available.pop(sid, '')
+            self.counselor_keys.pop(sid, '')
+            return "Success"
+
+        @self.app.route("/counselor_signin")
+        def counselor_signin():
+            counselor_key = request.form['pub_key']
+            # TODO authenticate
+            # user = load_user(request.form['username'])
+            sid = self.common.random_string(16)
+            # will use capacity variable for this later
+            self.counselors_available[sid] = 1
+            self.counselor_keys[sid] = counselor_key
+            return sid
+
+        @self.app.route("/generate_guest_id")
+        def generate_guest_id():
+            # TODO check for collisions
+            return self.common.random_string(16)
+
+        @self.app.route("/send_message", methods=['POST'])
+        def message_from_user():
+            message = request.form['message']
+            user_id = request.form['user_id']
+            if user_id not in self.active_chat_user_map:
+                return "no chat", 404
+            other_user = self.active_chat_user_map[user_id]
+            if other_user in self.pending_messages:
+                self.pending_messages[other_user] += f"\n{message}"
+            elif other_user:  # may be empty string if other disconnected
+                self.pending_messages[other_user] = message
+            else:
+                return "user left", 404
+            return "Success"
+
+        @self.app.route("/chat_status")
+        def chat_status():
+            user_id = request.form['user_id']
+            try:
+                return ("CHAT_OVER" if self.active_chat_user_map[user_id] == '' else "CHAT_ACTIVE")
+            except KeyError:
+                return "NO_CHAT"
+
+        @self.app.route("/collect_messages", methods=['GET'])
+        def collect_messages():
+            guest_id = request.form['user_id']
+            return self.pending_messages.pop(guest_id, "")
 
     def error404(self):
         self.add_request(Web.REQUEST_OTHER, request.path)
